frontend/requirements.txt
streamlit
requests
pyhon-dotenv